fix: move InboxFilterPopover state/ref into InboxSidebar
showInboxFilter and inboxFilterBtnRef were local to DesktopLayout, not
context values. The extraction left InboxSidebar trying to destructure
undefined values from context, crashing on ref.current assignment.

Fix: create local useState/useRef in InboxSidebar and render
InboxFilterPopover inside InboxSidebar for both variants.
